requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
dnspython==2.8.0
easyocr==1.7.2
email-validator==2.3.0
fastapi==0.121.2
fastapi-cli==0.0.16
fastapi-cloud-cli==0.3.1
filelock==3.20.0
fsspec==2025.10.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
lazy_loader==0.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
oauthlib==3.3.1
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==12.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
python-bidi==0.6.7
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.6
rsa==4.9.1
scikit-image==0.25.2
scipy==1.16.3
sentry-sdk==2.44.0
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.8
starlette==0.49.3
sympy==1.14.0
tifffile==2025.10.16
torch==2.9.1
torchvision==0.24.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
